rename files and move MonthCalendar component to /calendar
